+++ fsm/src/main/java/org/egov/fsm/web/model/user/User.java
@@ -9,7 +9,6 @@ import javax.validation.constraints.Pattern;
 import javax.validation.constraints.Size;
 
 import org.egov.common.contract.request.Role;
-import org.hibernate.validator.constraints.SafeHtml;
 import org.springframework.validation.annotation.Validated;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
@@ -32,83 +31,67 @@ public class User {
     private Long id;
 	 
     @Size(max=64)
-    @SafeHtml
     @JsonProperty("uuid")
     private String uuid;
 
     @Size(max=64)
-    @SafeHtml
     @JsonProperty("userName")
     private String userName;
     
 
     @Size(min=1,max=64)
-    @SafeHtml
     @Pattern(regexp = "^[a-zA-Z \\-'`\\.]*$", message = "Invalid name. Only alphabets and special characters -, ',`, .")
     @JsonProperty("name")
     private String name;
 
     @Size(max=64)
-    @SafeHtml
     @JsonProperty("password")
     private String password;
     
     @Size(max=64)
-    @SafeHtml
     @JsonProperty("mobileNumber")
     private String mobileNumber;
     
     @Size(max=64)
-    @SafeHtml
     @JsonProperty("tenantId")
     private String tenantId;
 
-    @SafeHtml
     @JsonProperty("salutation")
     private String salutation;
 
     @Size(max=128)
-    @SafeHtml
     @JsonProperty("emailId")
     private String emailId;
 
     @Size(max=50)
-    @SafeHtml
     @JsonProperty("altContactNumber")
     private String altContactNumber;
 
     @Size(max=10)
-    @SafeHtml
     @JsonProperty("pan")
     private String pan;
 
     @Pattern(regexp = "^[0-9]{12}$", message = "AdharNumber should be 12 digit number")
-    @SafeHtml
     @JsonProperty("aadhaarNumber")
     private String aadhaarNumber;
 
     @Size(max=300)
-    @SafeHtml
     @JsonProperty("permanentAddress")
     private String permanentAddress;
 
     @Size(max=300)
-    @SafeHtml
     @JsonProperty("permanentCity")
     private String permanentCity;
 
     @Size(max=10)
-    @SafeHtml
     @JsonProperty("permanentPinCode")
     private String permanentPincode;
 
     @Size(max=300)
-    @SafeHtml
     @JsonProperty("correspondenceCity")
     private String correspondenceCity;
 
     @Size(max=10)
-    @SafeHtml
     @JsonProperty("correspondencePinCode")
     private String correspondencePincode;
 
@@ -122,16 +105,13 @@ public class User {
     private Long pwdExpiryDate;
 
     @Size(max=16)
-    @SafeHtml
     @JsonProperty("locale")
     private String locale;
 
     @Size(max=50)
-    @SafeHtml
     @JsonProperty("type")
     private String type;
 
-    @SafeHtml
     @JsonProperty("signature")
     private String signature;
 
@@ -143,20 +123,16 @@ public class User {
     private List<Role> roles;
 
     @Size(max=32)
-    @SafeHtml
     @JsonProperty("bloodGroup")
     private String bloodGroup;
 
-    @SafeHtml
     @JsonProperty("identificationMark")
     private String identificationMark;
 
-    @SafeHtml
     @JsonProperty("photo")
     private String photo;
 
     @Size(max=64)
-    @SafeHtml
     @JsonProperty("createdBy")
     private String createdBy;
 
@@ -164,18 +140,15 @@ public class User {
     private Long createdDate;
 
     @Size(max=64)
-    @SafeHtml
     @JsonProperty("lastModifiedBy")
     private String lastModifiedBy;
 
     @JsonProperty("lastModifiedDate")
     private Long lastModifiedDate;
 
-    @SafeHtml
     @JsonProperty("otpReference")
     private String otpReference;
 
-    @SafeHtml
     @JsonProperty("gender")
     private String gender;